Some renaming & reorg, some forward progress
I didn't like the way the mental model for the project was moving. The
"RowManager" has become an "EventRouter". The stuff that was the EventBuffer
is now just the part of the Graph.

I'm standing by the decision not to have one buffer of events for each row,
instead of each graph. The data overhead is insignificant and it significantly
simplifies the code. (By insignificant, I mean the buffer takes up maybe 64
bytes per pong. It's never going to be a big array.) I need to remember to keep
trimming it after spans expire, but that's for Tomorrow Me.