Replaced table with divs.
I took some time to learn a thing or two about post-html2.0 page layout and
replaced the table with a series of nested grid divs. I still have an issue with
the height of the table not binding to the height of the page, and the canvases
being fixed at 300x150 (at least they're no longer being texture-magnified or
texture-minified anymore.) I'll deal with this later.