Near-rewrite of the UI side.
I've gone back and forth between using divs for layout and using tables, but in
the end, I've got a better chance of being able to allow the user to resize
specific columns of tables than if I use divs. That, and the tablular layout of
divs felt to arcane to me. The specification of my needs was just not as clear
as it is with tables.

The drawback here is that resizing the table cells isn't quite equating to an
exact size match. If I put an x-by-y canvas in a table cell and resize the cell
to x-by-y, I don't get a cell that properly contains - or even properly centers -
the canvas.

I'm leaving this latter problem to be solved later. I'll probably tackle it
before I start on the process of actually drawing graphs. There are still UI
decisions to be made before I settle on the data presentation, so I'll likely
get the cell sizing problem solved first.