Rewrite to depend upon (and add) user_icmp agent.
It sucks that there's no provision for icmp type 0 and 8 (echo response and
request, respectively) in most operating systems. Generally, ICMP controls how
a network is configured, so administrators don't want to give ICMP access to
unprivileged users, but the baby's been chucked with the bathwater. Rather than
having a way to do JUST those bits that are safe (as is done with TCP and UDP),
we're stuck with nothing.

This became a problem because this project needs a reliable way to do loss
tracking. In previous versions, this was done by estimating when the packet
SHOULD have been sent, then reporting it as missing when we expected it to have
arrived. This fails because of a design flaw in ping: the interpacket time is
measured by a timer that starts after the previous packet is sent. When that
timer expires, a new packet is sent. This should have been done by counting how
much time had elapsed since the session began and sending icmp_seq X when the
elapsed time had passed X*I, where I is the inter-packet interval time.

Because of this slop in the timing, I can only be accurate about when a packet
has DEFINITELY failed to arrive if packets after it have arrived. When there's
a long outage, we either can't report losses until the connection comes back up
(which would utterly suck for a tool that's supposed to be giving real-time net
behavior visualizations) or we have to guess which packets should have been
sent.  It turns out that this latter estimate is so error-prone that packets
were being reported as lost minutes before (or after) when they were actually
sent, which caused all sorts of hell in how the graph displayed data.

I toyed with the idea of writing a setuid script that ran:
  tcpdump -n -tt -v icmp[icmptype]==8 or icmp[icmptype]==0

... and filtered the output to report only the relevant data to the caller.
This could work, if written in C so it could be setuid. (Python and bash
scripts can't really be setuid.) I'm not certain that I should have abandoned
the C-calls-and-filters-tcpdump approach yet. But I've frequently lamented,
over the years, the crippling inconvenience that is icmp's status as a "raw
sockets only" function. Writing user_icmp is a step toward rectifying that.

Please understand that user_icmp must currently be run sudo. It still needs a
lot of work before I'm happy with the WAY it operates. Once that's done, I'll
rewrite the thing in C so it can be setuid.