Grooming of TODO list, Fix "Network is Unreachable"
When the socket realizes that the network in question (the ping target) is
unreachable, it throws an exception. I've chosen to just catch that exception
and pretend that the intended packet was actually sent. To the user, it may make
a difference WHY the packet couldn't make the round trip, but the specifics are
beyond the scope of user_icmp