Massive rewrite to eliminate the python backend.
There is still _a_ backend, and unfortunately it has to be run as root, but I
will get it rewritten entirely in C/++ so it can be compiled and setuid root.
This is a bit of a departure from the original intention of this project, but
with ICMP still languishing in the 1980s, there really wasn't much of a choice.