Added common base class for the graph and summary.
This was a significant refactor to fix a lot of the comms tangle that
had evolved over time. (There's still more.) This also moved handling
of common features (like rearranging geometry of canvases and containers)
to the base class.

get_max_displayed_rtt isn't quite as efficient, but there are bootstrapping
issues between get_max_displayed_rtt, the geometry of a canvas, and
add_event that needed some part of the process to give in and be the
first dependency.

Decisions about rebucketing have been simplified and the decision is now
made entirely by the layout manager.

I redraw the graph something like 10x/sec at the moment. Chrome isn't
running at more than about 10% cpu, so I'm good with that.

I still have this maddening issue with line 583, where the transiting
event that is being received doesn't have an entry. I suspect this is
happening when the packet in question was marked as lost and received
LONG afterward.